+++ src/realm/impl/transact_log.hpp
@@ -86,14 +86,20 @@ enum Instruction {
 
 class TransactLogStream {
 public:
+    virtual ~TransactLogStream()
+    {
+    }
+
     /// Ensure contiguous free space in the transaction log
     /// buffer. This method must update `out_free_begin`
     /// and `out_free_end` such that they refer to a chunk
     /// of free space whose size is at least \a n.
     ///
-    /// \param n The required amount of contiguous free space. Must be
+    /// \param size The required amount of contiguous free space. Must be
     /// small (probably not greater than 1024)
-    /// \param n Must be small (probably not greater than 1024)
+    /// \param out_free_begin must point to current write position which must be inside earlier
+    /// allocated area. Will be updated to point to new writing position.
+    /// \param out_free_end Will be updated to point to end of allocated area.
     virtual void transact_log_reserve(size_t size, char** out_free_begin, char** out_free_end) = 0;
 
     /// Copy the specified data into the transaction log buffer. This
@@ -112,8 +118,22 @@ public:
     void transact_log_reserve(size_t size, char** out_free_begin, char** out_free_end) override;
     void transact_log_append(const char* data, size_t size, char** out_free_begin, char** out_free_end) override;
 
-    const char* transact_log_data() const;
+    const char* data() const
+    {
+        return m_buffer.data();
+    }
 
+    char* data()
+    {
+        return m_buffer.data();
+    }
+
+    size_t size()
+    {
+        return m_buffer.size();
+    }
+
+private:
     util::Buffer<char> m_buffer;
 };
 
@@ -633,26 +653,21 @@ public:
 
 inline void TransactLogBufferStream::transact_log_reserve(size_t n, char** inout_new_begin, char** out_new_end)
 {
-    char* data = m_buffer.data();
-    REALM_ASSERT(*inout_new_begin >= data);
-    REALM_ASSERT(*inout_new_begin <= (data + m_buffer.size()));
-    size_t size = *inout_new_begin - data;
-    m_buffer.reserve_extra(size, n);
-    data = m_buffer.data(); // May have changed
-    *inout_new_begin = data + size;
-    *out_new_end = data + m_buffer.size();
+    char* buf = m_buffer.data();
+    REALM_ASSERT(*inout_new_begin >= buf);
+    REALM_ASSERT(*inout_new_begin <= (buf + m_buffer.size()));
+    size_t sz = *inout_new_begin - buf;
+    m_buffer.reserve_extra(sz, n);
+    buf = m_buffer.data(); // May have changed
+    *inout_new_begin = buf + sz;
+    *out_new_end = buf + m_buffer.size();
 }
 
-inline void TransactLogBufferStream::transact_log_append(const char* data, size_t size, char** out_new_begin,
+inline void TransactLogBufferStream::transact_log_append(const char* buf, size_t sz, char** out_new_begin,
                                                          char** out_new_end)
 {
-    transact_log_reserve(size, out_new_begin, out_new_end);
-    *out_new_begin = std::copy(data, data + size, *out_new_begin);
-}
-
-inline const char* TransactLogBufferStream::transact_log_data() const
-{
-    return m_buffer.data();
+    transact_log_reserve(sz, out_new_begin, out_new_end);
+    *out_new_begin = std::copy(buf, buf + sz, *out_new_begin);
 }
 
 inline TransactLogEncoder::TransactLogEncoder(TransactLogStream& stream)
@@ -2694,8 +2709,8 @@ private:
 
     size_t transact_log_size() const
     {
-        REALM_ASSERT_3(m_encoder.write_position(), >=, m_buffer.transact_log_data());
-        return m_encoder.write_position() - m_buffer.transact_log_data();
+        REALM_ASSERT_3(m_encoder.write_position(), >=, m_buffer.data());
+        return m_encoder.write_position() - m_buffer.data();
     }
 
     void append_instruction()
@@ -2745,7 +2760,7 @@ public:
         // push any pending select_table or select_descriptor into the buffer
         reverser.sync_table();
 
-        m_buffer = reverser.m_buffer.transact_log_data();
+        m_buffer = reverser.m_buffer.data();
         m_current = m_instr_order.size();
     }